// WADNR.Web/src/app/shared/generated/model/project-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectLookupItem { 
    ProjectID?: number;
    ProjectName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectLookupItemForm { 
    ProjectID?: FormControl<number>;
    ProjectName?: FormControl<string>;
}

export class ProjectLookupItemFormControls { 
    public static ProjectID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}